backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
numpy
httpx
python-dotenv
scipy
